- fix the following test case: __tmp1 = 1 tmp1 = 1 print - input() + 2 print __tmp1 print tmp1
The problem is that the compiler didn't keep track of the variable
names used by the p0 program, and therefore when generating temporary
variables names during the flattening phase, there were sometimes
collisions.  Fixed by keeping track of the variable names in the
StatementList object and generating a new temp. var name if one is
already used by the script.